+++ package/gluon-core/luasrc/lib/gluon/upgrade/180-outdoors
@@ -13,6 +13,7 @@ end
 
 local sysconfig = require 'gluon.sysconfig'
 local platform = require 'gluon.platform'
+local wireless = require 'gluon.wireless'
 
 local config = site.wifi5.outdoors('preset')
 local outdoor
@@ -22,7 +23,7 @@ if sysconfig.gluon_version then
 	outdoor = false
 elseif config == 'preset' then
 	-- enable outdoor mode through presets on new installs
-	outdoor = platform.is_outdoor_device() and platform.device_uses_11a(uci)
+	outdoor = platform.is_outdoor_device() and wireless.device_uses_11a(uci)
 else
 	-- enable/disable outdoor mode unconditionally on new installs
 	outdoor = config
