+++ package/gluon-core/luasrc/usr/lib/lua/gluon/platform.lua
@@ -1,7 +1,5 @@
 local platform_info = require 'platform_info'
 local util = require 'gluon.util'
-local wireless = require 'gluon.wireless'
-local unistd = require 'posix.unistd'
 
 
 local M = setmetatable({}, {
@@ -46,43 +44,6 @@ function M.is_outdoor_device()
 	end
 
 	return false
-end
-
-function M.device_supports_wpa3()
-	return unistd.access('/lib/gluon/features/wpa3')
-end
-
-function M.device_supports_mfp(uci)
-	local supports_mfp = true
-
-	if not M.device_supports_wpa3() then
-		return false
-	end
-
-	uci:foreach('wireless', 'wifi-device', function(radio)
-		local phy = wireless.find_phy(radio)
-		local phypath = '/sys/kernel/debug/ieee80211/' .. phy .. '/'
-
-		if not util.file_contains_line(phypath .. 'hwflags', 'MFP_CAPABLE') then
-			supports_mfp = false
-			return false
-		end
-	end)
-
-	return supports_mfp
-end
-
-function M.device_uses_11a(uci)
-	local ret = false
-
-	uci:foreach('wireless', 'wifi-device', function(radio)
-		if radio.hwmode == '11a' or radio.hwmode == '11na' then
-			ret = true
-			return false
-		end
-	end)
-
-	return ret
 end
 
 return M
